requirements.txt
Django==5.2.4
python-dateutil
Pillow
whitenoise